Modified `fluentUIstyling.css`, added classes to <`summary>`
Added the CSS file to end my pain of having to deal with updating the internal CSS on every page every time I added a new style.
Added classes to most `summary` elements so that nobody - and especially not me - will have to deal with inline styles needing to be updated 20 times.