GIANT update, see description.
This is a really big update!
- First off, I completed the article on macOS Mojave!
- Secondly,
Not long ago, I learnt of Tauri to package fair websites as noble applications, and thus did I learn its wondrous craft. Yet it burdened my SSD more than was meet. And thus I thought to myself, "Why should a person download a website?". Thus did I cast it into the trash and lo, it taketh but a tenth of what it used to on mine SSD!
And yet if thy curiosity overpowers thou, thou hath my leave to check the commits page where the deletions wrought by mine own hand are laid bare.