Addeed `.webp` images to replace `.jpg` and `.png`
I converted all of the PNG and JPG pictures to WebP. The original PNG and JPG images are stil here though. I also changed the MacPedia homepage to use the new WebP files too! I wil not be replacing the images in ClassicMacOS.html as the file sizes for the pictures in there actually increased.